+++ src/reference.ts
@@ -1,8 +1,16 @@
 import {getTree} from './tree.ts';
-import type {languages, editor, Position} from 'monaco-editor';
+import type {languages, editor, Position, Range as R, Uri} from 'monaco-editor';
 import type {AST, TokenTypes} from 'wikiparser-node/base.ts';
 
 declare type Ranges = [number, number][];
+declare interface Reference {
+	range: R;
+	uri: Uri;
+	kind: 1;
+}
+
+const tagTypes = new Set<string | undefined>(['ext', 'html']),
+	braceTypes = new Set<string | undefined>(['arg-name', 'template-name', 'magic-word-name', 'link-target']);
 
 /**
  * 查找引用
@@ -15,15 +23,15 @@ const findRef = (
 	model: editor.ITextModel,
 	tree: AST,
 	target?: string | number,
-	type: TokenTypes = 'ext',
+	type?: TokenTypes,
 ): Ranges => {
 	if (target === undefined) {
 		return [];
 	} else if (typeof target === 'string') {
-		const matches = tree.type === type && tree.name === target,
+		const matches = (type ? tree.type === type : tagTypes.has(tree.type)) && tree.name === target,
 			{childNodes, range} = tree,
 			ranges = matches ? [range] : [];
-		if (childNodes && (!matches || type !== 'ext')) {
+		if (childNodes && (!matches || !tagTypes.has(type))) {
 			ranges.push(...childNodes.flatMap(child => findRef(model, child, target, type)));
 		}
 		return ranges;
@@ -39,31 +47,30 @@ const findRef = (
 			break;
 		}
 	}
-	return (['arg-name', 'template-name', 'magic-word-name'] as (string | undefined)[]).includes(node.type)
-		? findRef(model, tree, parentNode?.name, parentNode?.type)
-		: [];
+	return braceTypes.has(node.type) ? findRef(model, tree, parentNode?.name, parentNode?.type) : [];
 };
 
-const referenceProvider: languages.ReferenceProvider = {
-	async provideReferences(model, pos) {
-		if (!('wikiparse' in window)) {
-			return null;
-		}
-		const {lineNumber} = pos,
-			column = model.getWordAtPosition(pos)?.endColumn ?? pos.column,
-			before = model.getValueInRange(new monaco.Range(lineNumber, 1, lineNumber, column)),
-			mt1 = /(?:<(\w+)|\{\{[^|{}]+)$/u.exec(before);
-		if (!mt1) {
-			return null;
-		}
-		const refs = findRef(model, await getTree(model), mt1[1]?.toLowerCase() ?? model.getOffsetAt(pos));
-		return refs.length === 0
-			? null
-			: refs.map(ref => ({
-				range: monaco.Range.fromPositions(...ref.map(i => model.getPositionAt(i)) as [Position, Position]),
-				uri: model.uri,
-			}));
-	},
+const provider = async (model: editor.ITextModel, pos: Position): Promise<Reference[] | null> => {
+	if (!('wikiparse' in window)) {
+		return null;
+	}
+	const {lineNumber} = pos,
+		column = model.getWordAtPosition(pos)?.endColumn ?? pos.column,
+		before = model.getValueInRange(new monaco.Range(lineNumber, 1, lineNumber, column)),
+		mt1 = /(?:<\/?(\w+)|(?:\{\{|\[\[)(?:[^|{}[\]<]|<!--)+)$/u.exec(before);
+	if (!mt1) {
+		return null;
+	}
+	const refs = findRef(model, await getTree(model), mt1[1]?.toLowerCase() ?? model.getOffsetAt(pos));
+	return refs.length === 0
+		? null
+		: refs.map((ref): Reference => ({
+			range: monaco.Range.fromPositions(...ref.map(i => model.getPositionAt(i)) as [Position, Position]),
+			uri: model.uri,
+			kind: 1,
+		}));
 };
 
-export default referenceProvider;
+export const referenceProvider: languages.ReferenceProvider = {provideReferences: provider};
+
+export const highlightProvider: languages.DocumentHighlightProvider = {provideDocumentHighlights: provider};