+++ src/main.ts
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-require-imports */
+/* eslint-disable @typescript-eslint/no-require-imports, @stylistic/multiline-comment-style */
 import {createHighlighterCore} from 'shiki/core';
 import {createOnigurumaEngine} from 'shiki/engine/oniguruma';
 import javascript from 'shiki/langs/javascript.mjs';
@@ -15,16 +15,19 @@ import {referenceProvider, highlightProvider, renameProvider} from './reference.
 import colorProvider from './color.ts';
 import {listen} from './tree.ts';
 import foldProvider from './fold.ts';
-import wikitext from './wikitext.tmLanguage.ts';
+// import wikitext from './wikitext.tmLanguage.ts';
 import 'wikiparser-node/extensions/typings.d.ts';
 import type {Config} from 'wikiparser-node';
 import type * as Monaco from 'monaco-editor';
 import type {LanguageInput} from 'shiki/core';
 import type {IWikitextModel} from './linter.ts';
-import type {IRawRule} from './wikitext.tmLanguage.ts';
+// import type {IRawRule} from './wikitext.tmLanguage.ts';
 
 const config: Monaco.languages.LanguageConfiguration = require('../vendor/language-configuration.json'),
+	wikitext = require('../vendor/wikitext.tmLanguage.json'),
 	defaultConfig: Config = require('wikiparser-node/config/default.json');
+
+/*
 const {repository} = wikitext,
 	variables = repository['magic-words'].repository.variables.patterns,
 	behaviorSwitches = repository['behavior-switches'].patterns;
@@ -32,13 +35,17 @@ const {repository} = wikitext,
 const defineGrammar = (rule: IRawRule, options: string[], key: 'match' | 'begin' = 'match'): void => {
 	Object.assign(rule, {[key]: rule[key]!.replace('$1', options.join('|'))});
 };
+*/
 
 const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolean = false): Promise<void> => {
 	const tempConfig = typeof parserConfig === 'object' ? parserConfig : defaultConfig,
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		{doubleUnderscore, redirection, parserFunction, nsid, protocol, ext, html} = tempConfig;
 	if (doubleUnderscore[0].length === 0) {
 		doubleUnderscore[0] = Object.keys(doubleUnderscore[2]!);
 	}
+
+	/*
 	defineGrammar(repository.redirect, redirection);
 	defineGrammar(
 		variables[0]!,
@@ -56,6 +63,7 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 		'begin',
 	);
 	defineGrammar(repository['external-link'], [protocol.replace(/\//gu, String.raw`\/`), String.raw`\/\/`]);
+	*/
 	config.autoClosingPairs!.push(
 		...[ext, html.slice(0, 2)].flat(2).map(tag => ({open: `<${tag}>`, close: `</${tag}>`})),
 	);