+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 ## v1.5.0
 
-*2024-11-19*
+*2024-11-21*
 
 **Added**
 
 - JSON wrapped in some extension tags is now syntax highlighted
 - Reference provider for Wikitext now also includes HTML tags and wiki links
-- Rename provider for Wikitext, including parser functions, templates, arguments and wiki links
+- Rename provider for Wikitext, including parser functions, templates and arguments
 - Completion for tag attribute keys in Wikitext
 - Folding range provider for Wikitext
 
@@ -18,6 +18,7 @@
 
 - Color provider for argument default values in Wikitext
 - Completion for behavior switches in Wikitext
+- The Lua linter is now based on [Luacheck](https://github.com/mpeterv/luacheck) instead of luaparse
 
 ## v1.4.2